agents/requirements.txt
python-dotenv
pydantic
openai
requests
flask
flask-cors
pymongo
weasyprint